+++ doc/changelog.md
@@ -5,7 +5,7 @@
  - Renamed the `Elements` class to `Ruitk`
  - Renamed the `elements.js` file to `ruitk.js`
  - Fixed a bug with `addElements` where it falsely said that it added negative elements. Eg: Added -55 out of 11 new elements
- - Added a option called `dependencys` to let RUITK any extra dependencys a element needs 
+ - Added a option called `dependencies` to let RUITK any extra dependencies a element needs 
 #### documentation Changes
  - Added a documentation page for security issues
  - Added documentation for the `markdown` element